[REF] point_of_sale: move uom overrides on the good place
the override of the uom and uom category where made in the product.py
file. So we move them in their own file.
