+++ addons/point_of_sale/models/product.py
@@ -95,16 +95,3 @@ class ProductProduct(models.Model):
             'suppliers': supplier_list,
             'variants': variant_list
         }
-
-
-class UomCateg(models.Model):
-    _inherit = 'uom.category'
-
-    is_pos_groupable = fields.Boolean(string='Group Products in POS',
-        help="Check if you want to group products of this category in point of sale orders")
-
-
-class Uom(models.Model):
-    _inherit = 'uom.uom'
-
-    is_pos_groupable = fields.Boolean(related='category_id.is_pos_groupable', readonly=False)
