+++ addons/point_of_sale/models/uom.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+from odoo import fields, models
+
+class UomCateg(models.Model):
+    _inherit = 'uom.category'
+
+    is_pos_groupable = fields.Boolean(string='Group Products in POS',
+        help="Check if you want to group products of this category in point of sale orders")
+
+
+class Uom(models.Model):
+    _inherit = 'uom.uom'
+
+    is_pos_groupable = fields.Boolean(related='category_id.is_pos_groupable', readonly=False)
